backend/requirements.txt
asgiref==3.8.1 ; python_version >= "3.9" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.9" and python_version < "4.0" and platform_python_implementation != "PyPy"
cryptography==42.0.5 ; python_version >= "3.9" and python_version < "4.0"
django-environ==0.11.2 ; python_version >= "3.9" and python_version < "4"
django-extensions==3.2.3 ; python_version >= "3.9" and python_version < "4.0"
django-filter==24.2 ; python_version >= "3.9" and python_version < "4.0"
django-rest-knox==4.2.0 ; python_version >= "3.9" and python_version < "4.0"
django==4.2.11 ; python_version >= "3.9" and python_version < "4.0"
djangorestframework==3.15.1 ; python_version >= "3.9" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.9" and python_version < "4.0" and platform_python_implementation != "PyPy"
sqlparse==0.5.0 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.11.0 ; python_version >= "3.9" and python_version < "3.11"
tzdata==2024.1 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
